sqlmock is only needed in tests
