+++ storage/sql/sql_test.go
@@ -7,8 +7,26 @@ package sql
 
 import (
 	"errors"
+
+	"github.com/DATA-DOG/go-sqlmock"
+	"github.com/ortuman/jackal/log"
+	"github.com/ortuman/jackal/pool"
 )
 
 var (
 	errMySQLStorage = errors.New("MySQL storage error")
 )
+
+// NewMock returns a mocked SQL storage instance.
+func NewMock() (*Storage, sqlmock.Sqlmock) {
+	var err error
+	var sqlMock sqlmock.Sqlmock
+	s := &Storage{
+		pool: pool.NewBufferPool(),
+	}
+	s.db, sqlMock, err = sqlmock.New()
+	if err != nil {
+		log.Fatalf("%v", err)
+	}
+	return s, sqlMock
+}
